New command: studio snabb rstudio
Runs the RStudio IDE (https://www.rstudio.com) with the snabbr libraries
loaded and the same versions of all R packages that the rest of Studio uses.

This is to make it easy to hack R code and know that it will run the
same under nix. R has a big ecosystem with lots of libraries being
updated etc so compatibility problems are common if you use different
versions for development and deployment.