backend: Add inspector (preview)
The inspector is a DWIM (do what I mean) interface to the Studio
backend.

Studio is able to translate many input formats (e.g. a Snabb process
snapshot) into many output formats (e.g. a visualization from R.) The
tricky part for an end-user is how to operate it: you probably don't
want to have to write a nix expression by hand every time you are
inspecting a file.

The solution here is to give the file to the inspector. It will
automatically try to process the file in a large number of different
ways, most of which will fail because the file won't be the right
type, and then it will give you all of the results that built successfully.

So then it is easy to operate the backend. You just pass it a
file/directory/url and then browser whatever it happens to spit out.

Currently the processing modules in the inspector are just examples
and not really a true interface to the Studio backend code of the
present and future.