tools/raptorjit: Support running RaptorJIT code
Simple Nix API for running RaptorJIT code:

  raptorjit.evalFile <luaSourceFile>:
    Evaluate a Lua source file. Provide its output and the trace
    data produced during execution.
  raptorjit.evalString <luaSource>:
    Helper function to call evalFile with code in string.

The standard behavior is to use a special 'auditlog' branch of
RaptorJIT that makes the VM debug symbols (DWARF) available and also
logs the JIT internal data structures (audit.log). So when you run
code with this API it will produce enough information for the UI to
visualize the JIT behavior.

Quick example:

    $ nix-build -E 'with import tools/raptorjit {}; evalString "for i = 1, 100 do end"'

    $ ls result
    audit.log  output.txt  raptorjit.dwo

Have to also include the profiler data soon...