nixpkgs: Add nixpkgs overlay with studio + pharo
Overlays are functions that transform the Nix packages collection.
Here I am using overlays to add/customize packages instead of forking
nixpkgs. I hope overlays will be an easy way to keep in sync with
upstream nixpkgs.

The 'studio' package is new and covers creating the Pharo VM for
Studio and running that under VNC (with help from ratpoison wm.)