Reorganize backend
The Nix backend code is all in backend/ now, migrated from nixpkgs/
and tools/.

The top-level studio directory now has a default.nix that returns
nixpkgs overlayed with the studio packages.

The backend/toplevel directory has a default.nix that returns the
studio packages only. This can be used as an entry point to the studio
API. For example:

    nix-build -I studio=$HOME/git/studio/backend/toplevel \
              -E 'with import <studio>; ...'

will run the code ... with the Studio API imported.