studio test: Extend timeout from 30sec to 10min
This is a little unfortunate. However, the current test script does
make a call to Nix to run a RaptorJIT program, which does require
getting/building raptorjit and all of its dependencies, so the 30s
time limit is not realistic.

Could be that the test needs to be scaled back a bit to run quicker?