frontend: Copy pharo VM to ./ with a name based on nix derivation
Previously the Pharo image was copied to ./pharo.image each time. This
was problematic because it can overwrite a precious file with that
name and different versions of the image can clobber each other if we
are running several copies of Studio at once.

This change implements a basic compromise of picking a name based on
the nix derivation path.