RaptorJIT: Simplify code by assuming that DWARF data is in audit.log
Studio had included some complicated Nix code for trying to obtain
suitable DWARF debug data for the virtual machine being inspected.
This is not needed with recent versions of RaptorJIT because the DWARF
data is included inline in the log file and can be extracted from
there.

Simple fixes also for loading an audit.log in isolation without
profiler data.