Frontend: Improve layout of RaptorJIT IR code
Layout the IR nodes as a tree but resolve conflicts (nodes with two
parents) by always choosing the parent that leads to the longest
ancestry path. This guarantees that a given node is always placed
below all of its ancestors but otherwise keeps it simple and regular.

This is currently only applied to the "head" part of a trace and not
the "loop." This is due to a quirk/bug of the layout engine that I
haven't gotten to the bottom of yet.