Frontend: Add support for RaptorJIT declname
The UI can now read a new 'declname' debug attribute that provides the
declared name of a Lua function. This is used to revamp the UI to make
descriptions of code executed in traces simpler.

Some other improvements coming along for the ride...

The improvements only kick in when using a RaptorJIT version with
declname support.