nix/precache.nix: Nix expression to build relevant nixpkgs
This file only depends on the nix/ directory and not the rest of
Studio. This is useful for Docker. Early in the Dockerfile we can add
nix/ and build the nixpkgs dependencies to get them cached. Then on a
later layer we can add the full studio/ source tree which is more
often modified forcing rebuilds.