Frontend: Switch to GToolkit beta release
Upgrade to a much newer and much-changed release of GToolkit.

Update the Pharo code to match (minimal porting.)

Update the Nix code to package the binary instead of building from
source (too hard due to Rust dependency for graphics.)

Build on a standard GToolkit image instead of a custom base one. We
can get away with this in the short term because GToolkit happens to
include all of the Smalltalk packages that Studio depends on. Once we
want to pull in more libraries we'll probably need to do that impurely
e.g. with nix-shell.